add night time calculation and some code refactoring

+++ lib/driver/db.dart
@@ -2,7 +2,10 @@ import 'dart:async';
 import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:uuid/uuid.dart';
-import 'models.dart';
+
+import 'package:web_logbook_mobile/models/models.dart';
+import 'package:web_logbook_mobile/driver/db_structure.dart';
+import 'package:web_logbook_mobile/helpers/helpers.dart';
 
 class DBProvider {
   DBProvider._();
@@ -17,76 +20,31 @@ class DBProvider {
     return _database;
   }
 
-  static const tableLogbook = '''
-    CREATE TABLE IF NOT EXISTS logbook (
-      uuid TEXT PRIMARY KEY,
-      date TEXT NOT NULL,
-      departure_place TEXT,
-      departure_time TEXT,
-      arrival_place TEXT,
-      arrival_time TEXT,
-      aircraft_model TEXT,
-      reg_name TEXT,
-      se_time TEXT,
-      me_time TEXT,
-      mcc_time TEXT,
-      total_time TEXT,
-      day_landings INTEGER,
-      night_landings INTEGER,
-      night_time TEXT,
-      ifr_time TEXT,
-      pic_time TEXT,
-      co_pilot_time TEXT,
-      dual_time TEXT,
-      instructor_time TEXT,
-      sim_type TEXT,
-      sim_time TEXT,
-      pic_name TEXT,
-      remarks TEXT,
-      update_time INTEGER
-    );
-
-    CREATE UNIQUE INDEX IF NOT EXISTS logbook_uuid ON logbook(uuid);
-  ''';
-
-  static const viewLogbook = '''
-    CREATE VIEW IF NOT EXISTS logbook_view
-    AS
-    SELECT uuid, date,
-      substr(date,7,4) || substr(date,4,2) || substr(date,0,3) as m_date,
-      departure_place, departure_time, arrival_place, arrival_time,
-      aircraft_model, reg_name, se_time, me_time, mcc_time, total_time,
-      iif(day_landings='',0,day_landings) as day_landings,
-      iif(night_landings='',0,night_landings) as night_landings,
-      night_time, ifr_time, pic_time, co_pilot_time, dual_time, instructor_time,
-      sim_type, sim_time, pic_name, remarks, update_time
-    FROM logbook;
-  ''';
-
-  static const deletedItems = '''
-    CREATE TABLE IF NOT EXISTS deleted_items (
-        uuid TEXT PRIMARY_KEY,
-        table_name TEXT NOT NULL,
-        delete_time TEXT NOT NULL
-    );
-  ''';
-
   /// Opens the sqlite3 DB, runs the sql queries to create
   /// a DB structure on first run and might some upgrade sql
   /// queries in future
   Future<Database> initDB() async {
     String path = join(await getDatabasesPath(), 'logbook.db');
 
-    return await openDatabase(
+    final database = await openDatabase(
       path,
-      version: 1,
+      version: 2,
       onCreate: (Database db, int version) async {
         await db.execute(tableLogbook);
         await db.execute(viewLogbook);
         await db.execute(deletedItems);
+        await db.execute(airportTable);
+      },
+      onUpgrade: (db, oldVersion, newVersion) async {
+        switch (oldVersion) {
+          case 1:
+            await db.execute(airportTable);
+        }
       },
       onOpen: (db) async {},
     );
+
+    return database;
   }
 
   /// Returns all flight records from the database and orders them by date
@@ -121,7 +79,7 @@ class DBProvider {
 
   /// Creates a new record in the logbook table with the provided
   /// flight record data [fr].
-  Future insertFlightRecord(FlightRecord fr) async {
+  Future<dynamic> insertFlightRecord(FlightRecord fr) async {
     final db = await database;
 
     return await db!.rawInsert('''INSERT INTO logbook
@@ -204,16 +162,18 @@ class DBProvider {
     ]);
   }
 
-  /// Deletes a [FlightRecord] with the specified [uuid] from the logbook database table.
-  /// The function also inserts a corresponding row into the deleted_items table, indicating
-  /// the deletion and the time it occurred.
-  Future deleteFlightRecord(String uuid) async {
+  /// Deletes a flight record
+  Future<dynamic> deleteFlightRecord(String uuid) async {
     final db = await database;
+    if (db == null) return 'cannot initialize connection to database';
 
-    await db!.rawDelete(
+    final res = await db.rawDelete(
       '''DELETE FROM logbook WHERE uuid = ?''',
       [uuid],
     );
+
+    if (res == 0) return 'no record with uuid $uuid found';
+
     await db.rawInsert(
       '''INSERT INTO deleted_items (uuid, table_name, delete_time)
         VALUES(?, ?, ?)''',
@@ -223,49 +183,7 @@ class DBProvider {
         getEpochTime(),
       ],
     );
-  }
 
-  /// A function that synchronizes a flight record with the local database.
-  /// Given a [FlightRecord] object [fr], this function checks if a flight record with
-  /// the same UUID already exists in the local database. If it does, and the
-  /// [updateTime] property of the existing record is older than that of [fr], then
-  /// [fr] is updated in the database. If the record doesn't exist in the database,
-  /// [fr] is inserted as a new record.
-  Future syncFlightRecord(FlightRecord fr) async {
-    final db = await database;
-
-    final row = await db!.rawQuery(
-      '''SELECT * FROM logbook WHERE uuid = ?''',
-      [fr.uuid],
-    );
-
-    if (row.isNotEmpty) {
-      final currentFR = FlightRecord.fromData(row[0]);
-      if (currentFR.updateTime < fr.updateTime) {
-        await updateFlightRecord(fr);
-      }
-    } else {
-      await insertFlightRecord(fr);
-    }
-  }
-
-  /// Deletes the item with the given UUID from the corresponding table
-  /// in the database.
-  Future syncDeletedItems(DeletedItem di) async {
-    final db = await database;
-
-    await db?.rawDelete(
-      '''DELETE FROM ${di.tableName} WHERE uuid = ?''',
-      [di.uuid],
-    );
-  }
-
-  /// Removed records from the deleted_items table once they are uploaded
-  /// to the main app. This logic might be changed in future, but for now
-  /// will keep like this
-  Future cleanDeletedItems() async {
-    final db = await database;
-
-    await db?.rawDelete('''DELETE FROM deleted_items''');
+    return null;
   }
 }
